Items card changes and fetch in Home component
